use usize instead of u8 everywhere and bool array in candidate cell

+++ src/constants.rs
@@ -1,3 +1,3 @@
-pub const REGION_SIZE: u8 = 3;
+pub const REGION_SIZE: usize = 3;
 
-pub const GRID_SIZE: u8 = REGION_SIZE * REGION_SIZE;
+pub const GRID_SIZE: usize = REGION_SIZE * REGION_SIZE;
